requirements.txt
Django==5.0.1
django-crispy-forms==2.1
django-environ==0.11.2
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9